Add a 'before_sanitize' option.
A function expected to return html that is run through the parser, is passed
true/false to indicate whether it's an update or not. My hope was that true/false
would indicate if it was going to be the submitted value.

redo of b198bc71cb5b0f93313cbfda08edcbc75b720166